LRQA-39865 Move setting the axis TestClassGroups into a parent class
